refactor: build external api modules
