+++ packages/perftool/.gitignore
@@ -1,0 +1,1 @@
+dist
